+++ rds-20140815/core/Models/CreateDdrInstanceRequest.cs
@@ -83,6 +83,12 @@ namespace AlibabaCloud.SDK.Rds20140815.Models
         [Validation(Required=false)]
         public string DBInstanceStorageType { get; set; }
 
+        /// <summary>
+        /// User-defined key ID for cloud disk encryption. Passing this parameter means turning on cloud disk encryption (it cannot be turned off after turning it on), and RoleARN needs to be passed in. You can view the key ID in the key management service console, or create a new key. For more information, see [Creating a Key](https://help.aliyun.com/document_detail/181610.html).
+        /// 
+        /// > - This parameter is only applicable to RDS SQL Server instances.
+        /// > - You can also not pass this parameter and only need to pass in RoleARN, which means setting the cloud disk encryption type of the instance to the RDS managed service key (Default Service CMK).
+        /// </summary>
         [NameInMap("EncryptionKey")]
         [Validation(Required=false)]
         public string EncryptionKey { get; set; }
@@ -211,6 +217,11 @@ namespace AlibabaCloud.SDK.Rds20140815.Models
         [Validation(Required=false)]
         public string RestoreType { get; set; }
 
+        /// <summary>
+        /// The global resource descriptor (ARN) of the RDS cloud service account authorized by the primary account to access the KMS permission. You can view the ARN information through the [CheckCloudResourceAuthorized](https://next.api.aliyun.com/document/Rds/2014-08-15/CheckCloudResourceAuthorized) API.
+        /// 
+        /// > This parameter is only available for RDS SQL Server instances.
+        /// </summary>
         [NameInMap("RoleARN")]
         [Validation(Required=false)]
         public string RoleARN { get; set; }